+++ sys/src/cmd/map/libmap/sinusoidal.c
@@ -1,7 +1,7 @@
 #include "map.h"
 
 int
-Xsinusoidal(struct place *place, float *x, float *y)
+Xsinusoidal(struct place *place, double *x, double *y)
 {
 	*x = - place->wlon.l * place->nlat.c;
 	*y = place->nlat.l;
